+++ index.html
@@ -39,12 +39,13 @@
         </section>
         <section class="press">
             <div class="inner">
-                <strong>전체 언론사</strong>
-                <span>내가 구독한 언론사</span>
+                <div class="tab-press">
+                    <span class="active">전체 언론사</span>
+                    <span>내가 구독한 언론사</span>
+                </div>
                 <div class="swiper">
                     <div class="swiper-container">
                         <div class="swiper-wrapper"></div>
-                        
                     </div><!--swiper-container-->
                     <button role="button" class="button prev"><span class="blind">이전</span></button>
                     <button role="button" class="button next"><span class="blind">다음</span></button>
@@ -53,6 +54,6 @@
             </div>
         </section>
     </main>
-    <script src="src/main.js"></script>
+    <script src="src/main.js" type="module"></script>
 </body>
 </html>
