+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "News Stand",
   "main": "src/crawling.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node src/crawling.js"
